Added ImageOperations class to handle image selection and type conversion.
chooseImage() - opens a file chooser that allows the user to pick .png or .jpg files, then converts the image to a BufferedImage.
imageToByteArray - converts a BufferedImage to an array of bytes for storing in a database.
imageToIcon - converts a BufferedImage to an icon for displaying on the GUI.

Also modified AddCustomer and SearchCustomer to use these functions.